refactor: 移除 smartctl -l ps3ssd 选项
删除未公开文档支持的 PS3 SSD 厂商健康日志功能，并将 ps3stor 机箱列表计数字段恢复为 U8。

+++ lib/dev_ps3stor.cpp
@@ -15,7 +15,7 @@ namespace smartmon {
 
 ps3stor_channel *ps3stor_channel::s_channel;
 
-ps3stor_errno ps3stor_channel::get_enclcount(unsigned hostid, uint16_t &enclcount)
+ps3stor_errno ps3stor_channel::get_enclcount(unsigned hostid, uint8_t &enclcount)
 {
   ps3stor_msg_info reqinfo{};
   reqinfo.length = sizeof(reqinfo);
